requirements.txt
mgz
ipython
trueskill
pyyaml
orjson